Separating helper from feature methods into base class
